requirements.txt
deepgram-sdk
deeptranslator
soundfile
pysrt
moviepy
requests
python-dotenv
youtube_dl